pyproject.toml
[tool]

[tool.poetry]
name = "quantide"
version = "0.1.0"
homepage = "https://github.com/zillionare/quantide"
description = "Skeleton project created by Python Project Wizard (ppw)."
authors = ["aaron yang <aaron_yang@jieyu.ai>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.13'
]

packages = [
    { include = "quantide" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"

black  = { version = "^22.3.0", optional = true}
isort  = { version = "5.10.1", optional = true}
flake8  = { version = "4.0.1", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest  = { version = "^7.0.1", optional = true}
pytest-cov  = { version = ">=7.0.0,<8.0.0", optional = true}
tox  = { version = "^3.24.5", optional = true}
mkdocs  = { version = "^1.2.3", optional = true}
mkdocs-include-markdown-plugin  = { version = "^3.2.3", optional = true}
mkdocs-material  = { version = "^8.1.11", optional = true}
mkdocs-material-extensions  = { version = "^1.0.3", optional = true}
twine  = { version = "^3.8.0", optional = true}
mkdocs-autorefs = {version = "^0.3.1", optional = true}
pre-commit = {version = "^2.17.0", optional = true}
toml = {version = "^0.10.2", optional = true}
livereload = {version = "^2.6.3", optional = true}
pyreadline = {version = "^2.1", optional = true}
mike = { version="^1.1.2", optional=true}
mypy = {version = "^1.5.1", optional = true}
apscheduler = "^3.10.4"
arrow = "^1.3.0"
cfg4py = "^0.12.1"
python-fasthtml = "^0.12.36"
bcrypt = "^5.0.0"
loguru = "^0.7.3"
aiosmtplib = "^5.0.0"
bidict = "^0.23.1"
tenacity = "^9.1.2"
sqlite-utils = "^3.39"
polars = "^1.36.1"
polars-talib = "^0.1.5"
pandas = "^2.3.3"
pyarrow = "^22.0.0"
monsterui = "^1.0.36"
redis = "^7.2.0"
cffi = ">=1.17.1"
msgpack = "^1.1.2"
tushare = "^1.4.25"
websockets = "^15.0.1"
[tool.poetry.extras]
test = [
    "pytest",
    "pytest-cov",
    "testcontainers"
    ]

dev = ["tox",
    "pre-commit",
    "twine",
    "toml",
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "mypy"
    ]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
    "mike",
    "setuptools",
    "pkginfo",
    "virtualenv"
    ]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.history
)/
'''
[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "pyarrow-stubs (>=20.0.0.20251215)",
    "freezegun (>=1.5.5,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
